+++ CodeChat_Server/tests/test_server.py
@@ -10,7 +10,6 @@
 # ----------------
 import asyncio
 import json
-from pathlib import Path
 import socketserver
 import subprocess
 
@@ -21,7 +20,6 @@ import requests
 # Local imports
 # -------------
 from CodeChat_Server.gen_py.CodeChat_Services.ttypes import RenderClientReturn
-from CodeChat_Server.__main__ import parse_args
 from CodeChat_Server.renderer import _render_CodeChat, _render_markdown, _render_ReST
 from CodeChat_Server.render_manager import (
     WEBSOCKET_PORT,
@@ -104,64 +102,6 @@ async def atest_5():
 
 def test_5(editor_plugin):
     asyncio.run(atest_5())
-
-
-# Command line parsing
-# --------------------
-def test_6():
-    assert vars(parse_args([])) == dict(
-        watch=[],
-        patterns=["*"],
-        ignore_patterns=[],
-        build=[],
-    )
-
-    assert vars(parse_args(["--watch"])) == dict(
-        watch=[str(Path(".").absolute())],
-        patterns=["*"],
-        ignore_patterns=[],
-        build=[],
-    )
-
-    assert vars(parse_args(["--watch", "1"])) == dict(
-        watch=["1"],
-        patterns=["*"],
-        ignore_patterns=[],
-        build=[],
-    )
-
-    assert vars(parse_args(["--watch", "--watch", "1"])) == dict(
-        watch=["1"],
-        patterns=["*"],
-        ignore_patterns=[],
-        build=[],
-    )
-
-    assert vars(
-        parse_args(
-            [
-                "--watch",
-                "1",
-                "--pattern",
-                "*.txt",
-                "*.rst",
-                "--ignore-pattern",
-                "*.html",
-            ]
-        )
-    ) == dict(
-        watch=["1"],
-        patterns=["*.txt", "*.rst"],
-        ignore_patterns=["*.html"],
-        build=[],
-    )
-
-    assert vars(parse_args(["--build", "5"])) == dict(
-        watch=[],
-        patterns=["*"],
-        ignore_patterns=[],
-        build=["5"],
-    )
 
 
 # Renderer tests